% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculateKineticsForEachGene}
\alias{calculateKineticsForEachGene}
\title{Calculate time-dependent alpha (transcription rate) and gamma (degradation rate) for each gene using observed gene expression of m_u, m_total}
\usage{
calculateKineticsForEachGene(
  observedData,
  boolCalculateInMinutes = TRUE,
  totalTimeOfOneCCInHours = 19.33,
  totalTimeOfOneCCInMinutes = 19.33 * 60,
  numberOfSplineAnchors = 20,
  boolSetBetaConstantBeforeAlphaEstimate = TRUE,
  ignoreOutliers = TRUE,
  boolUseCPM = TRUE,
  smallestLabelingTime = NULL
)
}
\arguments{
\item{boolCalculateInMinutes}{logic statement (default: TRUE)}

\item{totalTimeOfOneCCInHours}{numeric total cell cycle time in hours (default: 19.33)}

\item{totalTimeOfOneCCInMinutes}{numeric total cell cycle time in minutes (default: 19.33*60)}

\item{numberOfSplineAnchors}{number of spline anchors used for smoothing profiles (default: 20)}

\item{boolSetBetaConstantBeforeAlphaEstimate}{logic statement (default: TRUE)}

\item{ignoreOutliers}{logic statement (default: TRUE)}

\item{boolUseCPM}{logic statement (default: TRUE)}

\item{smallestLabelingTime}{a number the smallest metabolic labeling time (default: 15)}

\item{observedCounts}{data frame}
}
\value{
a data frame with kinetic rates of transcription (alpha) and degradation (gamma) for each gene at each cell cycle time
}
\description{
Calculate time-dependent alpha (transcription rate) and gamma (degradation rate) for each gene using observed gene expression of m_u, m_total
}
\examples{
data_rate <- calculateKineticsForEachGene(observedCounts, smallestLabelingTime = 15)

}
