% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateKinetics.R
\name{estimateKinetics}
\alias{estimateKinetics}
\title{Calculate kinetic rates for genes using multiple thread}
\usage{
estimateKinetics(smoothedObservedData, threadN = 8, numberOfSplineAnchors = 20)
}
\arguments{
\item{threadN}{an integer the number of thread for the calculation (default: 8)}

\item{observedData}{a data frame with gene expression of m_u, m_total, p_u and p_total in sorted cells.}

\item{numberofSplineAnchors}{an integer for the number of anchor points used for smooth the gene expression profiles along the cell cycle  (default: 20)}
}
\value{
a data frame with columns of gene, cc_time (cell cycle time), transcription rate (alpha), degradation rate (gamma).
}
\description{
Calculate kinetic rates for genes using multiple thread
}
\examples{
data_rates <- estimateKinetics(observedData = observed.data, threadN = 8, numberOfSplineAnchors = 20)



calculate kinetic rates for multiple genes using multiple thread
}
