% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInput.R
\name{getInput}
\alias{getInput}
\title{Formatting input data by combine gene counts with cell cycle time}
\usage{
getInput(
  rawCountsLabeledMature,
  rawCountsUnlabeledMature,
  rawCountsLabeledPrecursor,
  rawCountsUnlabeledPrecursor,
  cells,
  labelingTime
)
}
\arguments{
\item{rawCountsLabeledMature}{a data frame
count gene expression matrix of labeled mature RNAs, rownames are genes are colnames are cells.}

\item{rawCountsUnlabeledMature}{a data frame
count gene expression matrix of unlabeled mature RNAs, rownames are genes are colnames are cells.}

\item{rawCountsLabeledPrecursor}{a data frame
count gene expression matrix of unlabeled precursor RNAs, rownames are genes are colnames are cells.}

\item{rawCountsUnlabeledPrecursor}{a data frame
count gene expression matrix of unlabeled precursor RNAs, rownames are genes are colnames are cells.}

\item{labelingTime}{numeric the 4sU labeling time for the cells}

\item{cellInfo}{a data frame with columns:
cell_id (unique cell id)
cc_Time (cell cycle time in minutes)}
}
\value{
a data frame with columns:
   gene,
   cc_time,
   labeling_time,
   p_u,
   p_total,
   m_u,
   m_total,
   captureEfficiencyPerCell
}
\description{
Formatting input data by combine gene counts with cell cycle time
}
\examples{
observedData <- getInput(rawCounts = raw_counts, cellInfo=cells_info)


}
