% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPredictions.R
\name{getPredictions}
\alias{getPredictions}
\title{Calculate predicted gene expression based on estimated kinetic rates (only alpha and gamma)
using a simplified model.}
\usage{
getPredictions(dataRates, threadN = 8)
}
\arguments{
\item{dataRates}{a data frame with estimated time-dependent transcription (alpha) and degradation (gamma) rates}

\item{threadN}{an integer the number of thread that used for calculation (default: 8)}
}
\value{
a data frame with predicted gene expression of m_u and m_total for genes in sorted cells.
}
\description{
Calculate predicted gene expression based on estimated kinetic rates (only alpha and gamma)
using a simplified model.
}
\examples{
prediction <- getPredictions(dataRates, threadN = 8)

}
