% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPredictionSimplifiedModelForEachGene}
\alias{getPredictionSimplifiedModelForEachGene}
\title{Get predictions from estimated kinetic rates using simplified model}
\usage{
getPredictionSimplifiedModelForEachGene(
  parameterDF,
  boolCheckedInputCCTimeHasNoDuplicates = FALSE,
  boolCalculateInMinutes = TRUE,
  totalTimeOfOneCCInHours = 19.33,
  totalTimeOfOneCCInMinutes = 19.33 * 60,
  lengthToSimulatePerRoundInHours = 5,
  lengthToSimulatePerRoundInMinutes = 5 * 60,
  toleranceMature = 1,
  maximumNumberOfPeriodsToSimulate = 10,
  bytePrecisionForLargerNumbersToUse = 60
)
}
\arguments{
\item{parameterDF}{a data frame with transcription (alpha) and degradation (gamma) rate for each gene at each cell cycle time}

\item{boolCheckedInputCCTimeHasNoDuplicates}{logical statement (default: TRUE)}

\item{boolCalculateInMinutes}{logical statement (default: TRUE)}

\item{totalTimeOfOneCCInHours}{numeric total cell cycle time in hours (default: 19.33)}

\item{totalTimeOfOneCCInMinutes}{numeric total cell cycle time in minutes (default: 19.33*60)}

\item{lengthToSimulatePerRoundInHours}{numeric (default: 5)}

\item{lengthToSimulatePerRoundInMinutes}{numeric (default: 5*60)}

\item{toleranceMature}{numeric (default: 1)}

\item{maximumNumberOfPeriodsToSimulate}{numeric (default: 10)}

\item{bytePrecisionForLargerNumbersToUse}{numeric (default: 60)}
}
\value{
a data frame with predicted gene expression for both labeled and unlabeled mature RNAs at each cell cycle time.
}
\description{
Get predictions from estimated kinetic rates using simplified model
}
\examples{
data_predicted <- getPredictionSimplifiedModelForEachGene(parameterDF = data_rate)
p_u' = rep(NA, dim(parameterDF)[1]),
p_total' = rep(NA, dim(parameterDF)[1]),
p_u' = rep(NA, dim(parameterDF)[1]),
p_total' = rep(NA, dim(parameterDF)[1]),
p_u' = rep(NA, dim(parameterDF)[1]),
p_total' = rep(NA, dim(parameterDF)[1]),
p_u' = rep(NA, length(associatedtValuesThisLoop)),
p_total' = rep(NA, length(associatedtValuesThisLoop)),
p_u' = rep(NA, length(associatedtValuesThisLoop)),
p_total' = rep(NA, length(associatedtValuesThisLoop)),
}
