% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normalizeCountsToConcentration}
\alias{normalizeCountsToConcentration}
\title{Normalize gene expression count data to concentration}
\usage{
normalizeCountsToConcentration(rawData, umiPerCell = 1e+06)
}
\arguments{
\item{rawData}{a data frame of raw counts of 'm_u', [m_l', 'm_total', 'P_l', p_u', 'p_total' as columns, and 'captureEfficiencyPerCell' is also provide as a column}

\item{umiPerCell}{numeric molecules per cell (default: 1e6)}
}
\value{
data frame with gene counts normalized to concentrations
}
\description{
Normalize gene expression count data to concentration
}
\examples{
data.norm <- normalizeCountsToConcentration(rawData, umiPerCell = 1e6)

}
