% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRate.R
\name{plotMtotalAlphaGamma}
\alias{plotMtotalAlphaGamma}
\title{Plot the predicted gene expression, transcription (alpha) and degradation (gamma) rate profiles over the cell cycle time.}
\usage{
plotMtotalAlphaGamma(
  df.prediction = prediction,
  gene.to.plot = "PCNA",
  color.alpha = "#D95F02",
  color.gamma = "#1B9E77",
  color.expression = "#666666",
  show.phase.boundaries = TRUE,
  line.size = 1,
  font.size = 16
)
}
\arguments{
\item{df.prediction}{a data frame of the predicted gene expression, transcription (alpha) and degradation (gamma) with gene in sorted cells.}

\item{gene.to.plot}{character the name of the gene to plot (default: "PCNA")}

\item{color.alpha}{character the color used for transcription rate (default: "#D95F02")}

\item{color.gamma}{character the color used for degradation rate (default: "#1B9E77")}

\item{color.expression}{character the color used for predicted expression (default: "#666666")}

\item{show.phase.boundaries}{logical weather to plot the cell cycle phase boundaries in the x axis (default: TRUE)}

\item{line.size}{numeric the line size for the profile plot (default: 1)}

\item{font.size}{numeric the font size for the plot (default: 16)}
}
\value{

}
\description{
Plot the predicted gene expression, transcription (alpha) and degradation (gamma) rate profiles over the cell cycle time.
}
\examples{
plotMtotalAlphaGamma(df.prediction=df_predicted, gene.to.plot="PCNA")

}
